+++ orga/app/components/sco-organization-participant/header-actions.hbs
@@ -1,21 +1,10 @@
 <div class="organization-participant-list-page__header">
   <h1 class="page-title">{{t "pages.sco-organization-participants.title" count=@participantCount}}</h1>
   {{#if this.currentUser.isAdminInOrganization}}
-    {{#if @isLoading}}
-      <PixMessage @type="information" @withIcon="true">
-        {{t "pages.sco-organization-participants.actions.import-file.information"}}
-      </PixMessage>
-    {{/if}}
     <div class="organization-participant-list-page__import-students-button hide-on-mobile">
-      {{#if @isLoading}}
-        <PixButton @id="students-file-upload" @onChange={{@onImportStudents}} @isDisabled="true" @isLoading="true">
-          {{this.importButtonLabel}}
-        </PixButton>
-      {{else}}
-        <PixButtonUpload @id="students-file-upload" @onChange={{@onImportStudents}} accept={{this.acceptedFileType}}>
-          {{this.importButtonLabel}}
-        </PixButtonUpload>
-      {{/if}}
+      <PixButtonLink @route="authenticated.import-organization-participants">
+        {{t "pages.sco-organization-participants.actions.import-file"}}
+      </PixButtonLink>
     </div>
   {{/if}}
 </div>
