+++ orga/app/components/sco-organization-participant/header-actions.js
@@ -5,18 +5,4 @@ export default class ScoHeaderActions extends Component {
   @service currentUser;
   @service session;
   @service intl;
-
-  get acceptedFileType() {
-    if (this.currentUser.isAgriculture) {
-      return ['.csv'];
-    }
-    return ['.xml', '.zip'];
-  }
-
-  get importButtonLabel() {
-    const types = this.acceptedFileType.join(
-      this.intl.t('pages.sco-organization-participants.actions.import-file.file-type-separator'),
-    );
-    return this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types });
-  }
 }
