+++ orga/app/controllers/authenticated/import-organization-participants.js
@@ -19,7 +19,7 @@ export default class ImportController extends Controller {
   @tracked isLoading = false;
 
   @action
-  async importStudents(files) {
+  async importSupStudents(files) {
     const adapter = this.store.adapterFor('students-import');
     const organizationId = this.currentUser.organization.id;
 
@@ -34,6 +34,30 @@ export default class ImportController extends Controller {
     } finally {
       this.isLoading = false;
     }
+  }
+
+  @action
+  async importScoStudents(files) {
+    const adapter = this.store.adapterFor('students-import');
+    const organizationId = this.currentUser.organization.id;
+    const format = this.currentUser.isAgriculture ? 'csv' : 'xml';
+    const confirmBeforeClose = (event) => {
+      event.preventDefault();
+      return (event.returnValue = '');
+    };
+    window.addEventListener('beforeunload', confirmBeforeClose);
+    this.isLoading = true;
+    this.notifications.clearAll();
+    try {
+      await adapter.importStudentsSiecle(organizationId, files, format);
+      this.isLoading = false;
+      this.notifications.sendSuccess(this.intl.t('pages.organization-participants-import.global-success'));
+      this.router.transitionTo('authenticated.sco-organization-participants.list');
+    } catch (errorResponse) {
+      this.isLoading = false;
+      this._handleError(errorResponse);
+    }
+    window.removeEventListener('beforeunload', confirmBeforeClose);
   }
 
   @action
@@ -101,5 +125,28 @@ export default class ImportController extends Controller {
         onClick: () => window.open(this.intl.t('common.help-form'), '_blank'),
       });
     }
+  }
+
+  _handleError(errorResponse) {
+    const globalErrorMessage = this.intl.t('pages.sco-organization-participants.import.global-error', {
+      htmlSafe: true,
+    });
+    if (!errorResponse.errors) {
+      return this.notifications.sendError(globalErrorMessage, {
+        onClick: () => window.open(this.intl.t('common.help-form'), '_blank'),
+      });
+    }
+
+    errorResponse.errors.forEach((error) => {
+      if (['422', '412', '413'].includes(error.status)) {
+        const message = this.errorMessages.getErrorMessage(error.code, error.meta) || error.detail;
+        return this.notifications.sendError(
+          this.intl.t('pages.sco-organization-participants.import.error-wrapper', { message, htmlSafe: true }),
+        );
+      }
+      return this.notifications.sendError(globalErrorMessage, {
+        onClick: () => window.open(this.intl.t('common.help-form'), '_blank'),
+      });
+    });
   }
 }
