+++ orga/tests/integration/components/sco-organization-participant/header-actions_test.js
@@ -1,10 +1,8 @@
 import { module, test } from 'qunit';
 import { render } from '@1024pix/ember-testing-library';
 import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
-import { triggerEvent } from '@ember/test-helpers';
 import Service from '@ember/service';
 import { hbs } from 'ember-cli-htmlbars';
-import sinon from 'sinon';
 
 module('Integration | Component | ScoOrganizationParticipant::HeaderActions', function (hooks) {
   setupIntlRenderingTest(hooks);
@@ -37,117 +35,45 @@ module('Integration | Component | ScoOrganizationParticipant::HeaderActions', fu
     });
   });
 
-  module('user rights', () => {
-    module('when user is admin in organization', () => {
-      module('when organization is SCO', (hooks) => {
-        let screen;
-        let importStudentsSpyStub;
-        hooks.beforeEach(async function () {
-          class CurrentUserStub extends Service {
-            isAdminInOrganization = true;
-          }
-          this.owner.register('service:current-user', CurrentUserStub);
-          importStudentsSpyStub = sinon.stub();
-          this.set('importStudentsSpy', importStudentsSpyStub);
-        });
-
-        test('it trigger importStudentsSpy when clicking on the import button', async function (assert) {
-          screen = await render(
-            hbs`<ScoOrganizationParticipant::HeaderActions @onImportStudents={{this.importStudentsSpy}} @isLoading={{false}} />`,
-          );
-
-          const file = new Blob(['foo'], { type: 'valid-file' });
-
-          const input = screen.getByLabelText(
-            this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types: '.xml ou .zip' }),
-          );
-          await triggerEvent(input, 'change', { files: [file] });
-
-          assert.ok(importStudentsSpyStub.called);
-        });
-
-        test('the import button should be in loading state when importing', async function (assert) {
-          screen = await render(
-            hbs`<ScoOrganizationParticipant::HeaderActions @onImportStudents={{this.importStudentsSpy}} @isLoading={{true}} />`,
-          );
-
-          assert.dom(screen.getByRole('button', { hidden: true })).exists();
-        });
-
-        test('a message should be display when importing ', async function (assert) {
-          screen = await render(
-            hbs`<ScoOrganizationParticipant::HeaderActions @onImportStudents={{this.importStudentsSpy}} @isLoading={{true}} />`,
-          );
-          assert
-            .dom(
-              await screen.findByText(
-                this.intl.t('pages.sco-organization-participants.actions.import-file.information'),
-              ),
-            )
-            .exists();
-        });
-      });
-
-      module('when organization is SCO and tagged as Agriculture and CFA', (hooks) => {
-        let screen;
-        hooks.beforeEach(async function () {
-          class CurrentUserStub extends Service {
-            isAdminInOrganization = true;
-            isAgriculture = true;
-            organization = {};
-            prescriber = {
-              lang: 'fr',
-            };
-          }
-
-          this.set('importStudentsSpy', () => {});
-          this.owner.register('service:current-user', CurrentUserStub);
-        });
-
-        test('the import button should be in loading state when importing', async function (assert) {
-          screen = await render(
-            hbs`<ScoOrganizationParticipant::HeaderActions @onImportStudents={{this.importStudentsSpy}} @isLoading={{true}} />`,
-          );
-
-          assert.dom(screen.getByRole('button', { hidden: true })).exists();
-        });
-
-        test('a message should be display when importing ', async function (assert) {
-          screen = await render(
-            hbs`<ScoOrganizationParticipant::HeaderActions @onImportStudents={{this.importStudentsSpy}} @isLoading={{true}} />`,
-          );
-          assert
-            .dom(
-              await screen.findByText(
-                this.intl.t('pages.sco-organization-participants.actions.import-file.information'),
-              ),
-            )
-            .exists();
-        });
-      });
-    });
-
-    module('when user is not admin in organization', () => {
-      test('it should not display import button', async function (assert) {
+  module('when user is admin in organization', () => {
+    module('when organization is SCO', (hooks) => {
+      hooks.beforeEach(function () {
         class CurrentUserStub extends Service {
-          isAdminInOrganization = false;
+          isAdminInOrganization = true;
         }
         this.owner.register('service:current-user', CurrentUserStub);
-        const screen = await render(hbs`<ScoOrganizationParticipant::HeaderActions />`);
-
-        assert.strictEqual(
-          screen.queryByLabelText(
-            this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types: '.csv' }),
-          ),
-          null,
-        );
-        assert.strictEqual(
-          screen.queryByLabelText(
-            this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types: '.xml ou .zip' }),
-          ),
-          null,
-        );
       });
+
+      test('it displays the import button', async function (assert) {
+        // when
+        await render(hbs`<ScoOrganizationParticipant::HeaderActions />`);
+
+        // then
+        assert.contains('Importer');
+      });
+    });
+  });
+
+  module('when user is not admin in organization', () => {
+    test('it should not display import button', async function (assert) {
+      class CurrentUserStub extends Service {
+        isAdminInOrganization = false;
+      }
+      this.owner.register('service:current-user', CurrentUserStub);
+      const screen = await render(hbs`<ScoOrganizationParticipant::HeaderActions />`);
+
+      assert.strictEqual(
+        screen.queryByLabelText(
+          this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types: '.csv' }),
+        ),
+        null,
+      );
+      assert.strictEqual(
+        screen.queryByLabelText(
+          this.intl.t('pages.sco-organization-participants.actions.import-file.label', { types: '.xml ou .zip' }),
+        ),
+        null,
+      );
     });
   });
 });
