+++ orga/translations/fr.json
@@ -957,11 +957,7 @@
         "reset-password-button": "Réinitialiser les mots de passe des élèves sélectionnés"
       },
       "actions": {
-        "import-file": {
-          "file-type-separator": " ou ",
-          "information": "L'import de votre fichier peut prendre quelques minutes. Certaines fonctionnalités sont temporairement inaccessibles.",
-          "label": "Importer ({types})"
-        },
+        "import-file": "Importer",
         "manage-account": "Gérer le compte",
         "show-actions": "Afficher les actions"
       },
